Rename excercise.js
